fix merge conflicts and add in logger statement to test_model_package
