+++ apis/v1alpha1/ack-generate-metadata.yaml
@@ -1,14 +1,14 @@
 ack_generate_info:
-  build_date: "2021-07-20T03:49:37Z"
+  build_date: "2021-07-20T22:57:41Z"
   build_hash: a8283f525037c5b9c42a4c3c723ba70e86678e17
   go_version: go1.16.4 darwin/amd64
   version: v0.6.0
-api_directory_checksum: 5557bf46f998baa97ec7f8fa7a29a00a4fe3124e
+api_directory_checksum: 298cfa6978039dcd7fdc6140e485e5ef090d6a01
 api_version: v1alpha1
 aws_sdk_go_version: ""
 generator_config_info:
-  file_checksum: dd0f056f6c1fa4789399cce1f53e6c3d6fce43a6
+  file_checksum: 79ae25a0d8659681207c3352caf4c1ba2b80e0f8
   original_file_name: generator.yaml
 last_modification:
   reason: API generation
-  timestamp: 2021-07-20 03:49:47.126715 +0000 UTC
+  timestamp: 2021-07-20 22:57:52.01637 +0000 UTC
