+++ pkg/resource/hyper_parameter_tuning_job/sdk.go
@@ -1652,7 +1652,7 @@ func (rm *resourceManager) sdkDelete(
 	// resource Unmanaged
 	latestStatus := r.ko.Status.HyperParameterTuningJobStatus
 	if latestStatus != nil && *latestStatus != svcsdk.HyperParameterTuningJobStatusInProgress {
-		return nil
+		return nil, nil
 	}
 	input, err := rm.newDeleteRequestPayload(r)
 	if err != nil {
