+++ pkg/resource/monitoring_schedule/sdk.go
@@ -943,7 +943,7 @@ func (rm *resourceManager) sdkDelete(
 	// specialized logic to check if modification is allowed
 	err = rm.statusAllowUpdates(ctx, r)
 	if err != nil {
-		return err
+		return nil, err
 	}
 	input, err := rm.newDeleteRequestPayload(r)
 	if err != nil {
