+++ pkg/resource/transform_job/sdk.go
@@ -454,7 +454,7 @@ func (rm *resourceManager) sdkDelete(
 	// resource Unmanaged
 	latestStatus := r.ko.Status.TransformJobStatus
 	if latestStatus != nil && *latestStatus != svcsdk.TransformJobStatusInProgress {
-		return nil
+		return nil, nil
 	}
 	input, err := rm.newDeleteRequestPayload(r)
 	if err != nil {
