+++ templates/common/sdk_delete_post_request.go.tpl
@@ -1,8 +1,8 @@
     if err == nil {
             if _, err := rm.sdkFind(ctx, r); err != ackerr.NotFound {
                 if err != nil {
-                    return err
+                    return nil, err
                 }
-                return requeueWaitWhileDeleting
+                return nil, requeueWaitWhileDeleting
             }
     }
