+++ templates/common/sdk_delete_pre_build_request.go.tpl
@@ -1,3 +1,3 @@
     if err = rm.requeueUntilCanModify(ctx, r); err != nil {
-            return err
+            return nil, err
         }
