+++ templates/hyper_parameter_tuning_job/sdk_delete_pre_build_request.go.tpl
@@ -2,5 +2,5 @@
 	// resource Unmanaged
 	latestStatus := r.ko.Status.HyperParameterTuningJobStatus
 	if latestStatus != nil && *latestStatus != svcsdk.HyperParameterTuningJobStatusInProgress {
-		return nil
+		return nil, nil
 	}
