+++ templates/model_package/sdk_delete_post_build_request.go.tpl
@@ -2,6 +2,6 @@
         if r.ko.Status.ACKResourceMetadata != nil && r.ko.Status.ACKResourceMetadata.ARN != nil {
             input.SetModelPackageName(string(*r.ko.Status.ACKResourceMetadata.ARN))
         } else {
-            return ackerr.NotFound
+            return nil, ackerr.NotFound
         }
     }
