--- templates/model_package/sdk_delete_pre_build_request.go.tpl
@@ -1,10 +1,0 @@
-	if isModelPackageDeleting(r) {
-        msg := "ModelPackage is currently being deleted"
-		setSyncedCondition(r, corev1.ConditionFalse, &msg, nil)
-		return requeueWaitWhileDeleting
-	}
-    if isModelPackageInProgress(r) {
-        msg := "ModelPackage is currently in progress"
-		setSyncedCondition(r, corev1.ConditionFalse, &msg, nil)
-		return requeueWaitWhileInProgress
-	}
