--- templates/model_package/sdk_read_one_post_set_output.go.tpl
@@ -1,11 +1,0 @@
-    if isModelPackagePending(&resource{ko}) {
-            msg := "ModelPackage is currently pending"
-		    setSyncedCondition(&resource{ko}, corev1.ConditionFalse, &msg, nil)
-            return &resource{ko}, requeueWaitWhilePending
-    }
-    if isModelPackageInProgress(&resource{ko}) {
-            msg := "ModelPackage is currently in progress"
-		    setSyncedCondition(&resource{ko}, corev1.ConditionFalse, &msg, nil)
-            return &resource{ko}, requeueWaitWhileInProgress
-    }
-    ModelPackageCustomSetOutput(&resource{ko})
