--- templates/model_package_group/sdk_read_one_post_set_output.go.tpl
@@ -1,11 +1,0 @@
-    if isModelPackageGroupPending(&resource{ko}) {
-        msg := "ModelPackageGroup is currently pending"
-        setSyncedCondition(&resource{ko}, corev1.ConditionFalse, &msg, nil)
-        return &resource{ko}, requeueWaitWhilePending
-    }
-    if isModelPackageGroupInProgress(&resource{ko}) {
-        msg := "ModelPackageGroup is currently in progress"
-        setSyncedCondition(&resource{ko}, corev1.ConditionFalse, &msg, nil)
-        return &resource{ko}, requeueWaitWhileInProgress
-    }
-    ModelPackageGroupCustomSetOutput(&resource{ko})
