+++ templates/monitoring_schedule/sdk_delete_pre_build_request.go.tpl
@@ -1,5 +1,5 @@
     // specialized logic to check if modification is allowed
     err = rm.statusAllowUpdates(ctx, r)
     if err != nil {
-        return err
+        return nil, err
     }
