+++ templates/training_job/sdk_delete_pre_build_request.go.tpl
@@ -2,5 +2,5 @@
 	// resource Unmanaged
 	latestStatus := r.ko.Status.TrainingJobStatus
 	if latestStatus != nil && *latestStatus != svcsdk.TrainingJobStatusInProgress {
-		return nil
+		return nil, nil
 	}
