+++ templates/transform_job/sdk_delete_pre_build_request.go.tpl
@@ -2,5 +2,5 @@
 	// resource Unmanaged
 	latestStatus := r.ko.Status.TransformJobStatus
 	if latestStatus != nil && *latestStatus != svcsdk.TransformJobStatusInProgress {
-		return nil
+		return nil, nil
 	}
